@86erg4ek2 #removed console logs, comments. added utils file
